pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toio-mcp"
version = "0.1.0"
description = "MCP server for toio Core Cube"
readme = "README.md"
authors = [
    {name = "Akihiro Komori", email = "akihiro.komori+toiomcp@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "toio.py",
    "uvicorn",
    "typer",
    "rich",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
toio-mcp = "toio_mcp.__main__:main"

[tool.setuptools]
packages = ["toio_mcp", "toio_mcp.tools"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
